add terminate function

+++ create_ec2_instance.py
@@ -7,16 +7,17 @@ import os
 
 class create_ec2_instances(object):
 	def __init__(self, 
-				aws_access_key_id = None,
+				aws_access_key_id = None, 
 				aws_secret_access_key = None, 
 				region_name = None,
 				ImageId = 'ami-c7eaf7b8', 
 				MaxCount = 1, 
 				MinCount = 1, 
 				InstanceType = 't2.micro', 
-				time_out = 30,):
+				time_out = 30,
+				user_data = None):
 		try: 
-			if not aws_access_key_id and not aws_secret_access_key and not region_name:
+			if aws_access_key_id and aws_secret_access_key and region_name:
 				self.session = boto3.Session(aws_access_key_id = aws_access_key_id,
 											  aws_secret_access_key = aws_secret_access_key,
 											  region_name = region_name)
@@ -32,13 +33,15 @@ class create_ec2_instances(object):
 		self.MinCount = MinCount
 		self.InstanceType = InstanceType
 		self.time_out = time_out
+		self.user_data = user_data
 		self.public_ip_address_list, self.id_list = self.create_running_instance()
 
 	def create_running_instance(self):
 		instances = self.ec2.create_instances(ImageId = self.ImageId,
 												MinCount = self.MinCount,
 												MaxCount = self.MaxCount,
-												InstanceType= self.InstanceType)
+												InstanceType= self.InstanceType,
+												UserData = self.user_data)
 
 		running_instance_num = 0
 		id_list = [ins.id for ins in instances]
@@ -63,24 +66,51 @@ class create_ec2_instances(object):
 
 		print("We created ", running_instance_num, "running instances")
 
-		return self.save_info(filtered_instances)		
+		return self.save_info(filtered_instances)	
 		# return public_ip_addresses, instances_id
+
+	# def save_info(self, filtered_instances):
+	# 	for instance in filtered_instances:
+	# 		print(instance.id, "public_ip_address is ", instance.public_ip_address)
+	# 	public_ip_addresses = [instance.public_ip_address for instance in filtered_instances]
+	# 	instances_id = [instance.id for instance in filtered_instances]
+	# 	instances_info = { 
+	# 					"public_ip_addresses": public_ip_addresses,
+	# 					"instances_id": instances_id
+	# 					}
+
+	# 	if not os.path.isdir("./tmp/"):
+	# 		os.mkdir("./tmp")
+	# 	with open("./tmp/instances_info.json",'w') as outfile:
+	# 		json.dump(instances_info, outfile, indent = 4 , sort_keys = True)
+	# 	return public_ip_addresses, instances_id
 
 	def save_info(self, filtered_instances):
 		for instance in filtered_instances:
-			print(instance.id, "public_ip_address is ", instance.public_ip_address)
+			print("instance id is ", instance.id, "public_ip_address is ", instance.public_ip_address)
 		public_ip_addresses = [instance.public_ip_address for instance in filtered_instances]
-		instances_id = [instance.id for instance in filtered_instances]
-		instances_info = { 
-						"public_ip_addresses": public_ip_addresses,
-						"instances_id": instances_id
-						}
+		instances_ids = [instance.id for instance in filtered_instances]
 
 		if not os.path.isdir("./tmp/"):
 			os.mkdir("./tmp")
-		with open("./tmp/instances_info.json",'w') as outfile:
-			json.dump(instances_info, outfile, indent = 4 , sort_keys = True)
-		return public_ip_addresses, instances_id
+		if not os.path.exists("./tmp/instances_info.json"):
+			instances_info = { 
+							"public_ip_addresses": public_ip_addresses,
+							"instances_id": instances_ids
+							}
+			with open("./tmp/instances_info.json",'w') as outfile:
+				json.dump(instances_info, outfile, indent = 4 , sort_keys = True)
+		else:
+			with open("./tmp/instances_info.json",'r') as infile:
+				instances_info = json.load(infile)
+				for public_ip_address in public_ip_addresses:
+					instances_info["public_ip_addresses"].append(public_ip_address)
+				for instance_id in instances_ids:
+					instances_info["instances_id"].append(instance_id)
+			with open("./tmp/instances_info.json",'w') as outfile:
+				json.dump(instances_info, outfile, indent = 4, sort_keys = True)
+		return public_ip_addresses, instances_ids
+
 
 	def terminate(self):
 		if (self.id_list == []):
@@ -102,7 +132,7 @@ class create_ec2_instances(object):
 		return self.id_list
 
 if __name__=="__main__":
-	ImageId = 'ami-0f793c9a57200fbb6'
+	ImageId = 'ami-095f89c443d0824b5'
 	instances = create_ec2_instances(ImageId = ImageId, MaxCount = 2, MinCount = 1, InstanceType = 'm4.xlarge')
 	print(instances.get_instance_public_ip())
 	print(instances.get_instance_id())
